# backend/openapi.yml
openapi: 3.0.3
info:
  title: STGY API
  version: 1.0.0
  description: |
    STGY SNS backend API (Express/TypeScript)
    - Cookie authentication: session_id
servers:
  - url: http://localhost:3001
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    StorageObjectMetadata:
      type: object
      properties:
        bucket: { type: string }
        key: { type: string }
        size: { type: integer }
        etag: { type: string, nullable: true }
        lastModified: { type: string, format: date-time, nullable: true }
        storageClass: { type: string, nullable: true }
        contentType: { type: string, nullable: true }
        metadata:
          type: object
          nullable: true
          additionalProperties: { type: string }
      required: [bucket, key, size]
...
    TrackObject:
      allOf:
        - $ref: "#/components/schemas/StorageObjectMetadata"
        - type: object
          properties:
            publicUrl: { type: string }
            previewKey: { type: string }
            previewUrl: { type: string }
          required: [publicUrl, previewKey, previewUrl]
...
  responses:
    Error400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
paths:
...
  /media/{userId}/tracks:
    get:
      summary: List user's tracks
      security:
        - sessionCookie: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
        - name: after
          in: query
          description: Exclusive object-key cursor; requires offset=0
          schema: { type: string }
      responses:
        "200":
          description: List of track master objects with preview URLs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackObject"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
...
  /media/{userId}/tracks/{path}:
    get:
      summary: Download a specific track master or preview
      description: |
        NOTE: {path} may contain slashes (server treats it as a catch-all).
        Masters are stored under masters/ and previews under previews/.
      security:
        - sessionCookie: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
        - name: path
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Track bytes
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
            application/gzip:
              schema: { type: string, format: binary }
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
    delete:
      summary: Delete a specific track master and its preview
      description: |
        NOTE: {path} may contain slashes (server treats it as a catch-all).
        Delete the master path; the derived preview object is deleted together.
      security:
        - sessionCookie: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
        - name: path
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  result: { type: string }
                required: [result]
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
...